Do not stall when using probe_count=1
When `probe_count = 1`, the `probe_group`'s `last_delay` will be set to the `delay` of this probe.
Now, when adding a "new" probe to discover the next hop to the `probe_group`, `next_delay` and `last_delay` will be identical, i.e. `next_delay - last_delay == 0`.
Thus `update_time` would actually disarm the timer, which causes the new probe to never be scheduled.